api: fix get_payment_info and get_refund_info
We weren't populating all values correctly when looking up
payment and refund information because the transformation code
was relying on the associated transaction (which doesn't exist
when going to PayPal).

We don't really need to go back to PayPal actually - just use the cached
version.

Note that the amount returned for refunds is positive now, which is
consistent with how it is stored.

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>